Refactor code for clarity
Grateful Drop" is a heartfelt NFT created to thank Optimism for the generous Airdrop 5 reward. This token symbolizes our appreciation for the support that empowers us to keep building and learning in the vibrant world of blockchain. With each unique piece, we celebrate the journey of growth and collaboration, inspiring us to embrace new opportunities and foster innovation together. Let this NFT be a reminder of our commitment to progress and the shared vision that drives us forward

Signed-off-by: YACCETH <116510143+YACCETH@users.noreply.github.com>